+++ jsonrpc/debug_test.go
@@ -5,20 +5,22 @@ import (
 
 	"github.com/cloudwalk/ethgo/testutil"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestDebug_TraceTransaction(t *testing.T) {
-	s := testutil.NewTestServer(t, nil)
-	defer s.Close()
-
+	s := testutil.NewTestServer(t)
 	c, _ := NewClient(s.HTTPAddr())
 
 	cc := &testutil.Contract{}
 	cc.AddEvent(testutil.NewEvent("A").Add("address", true))
 	cc.EmitEvent("setA", "A", addr0.String())
 
-	_, addr := s.DeployContract(cc)
-	r := s.TxnTo(addr, "setA2")
+	_, addr, err := s.DeployContract(cc)
+	require.NoError(t, err)
+
+	r, err := s.TxnTo(addr, "setA2")
+	require.NoError(t, err)
 
 	trace, err := c.Debug().TraceTransaction(r.TransactionHash)
 	assert.NoError(t, err)