+++ CHANGELOG.md
@@ -1,5 +1,18 @@
+# 0.1.4 (Unreleased)
 
-# 0.1.2 (Unreleased)
+- fix: Parse `nonce` and `mixHash` on `Block` [[GH-228](https://github.com/cloudwalk/ethgo/issues/228)]
+- feat: `abi` decodes function string in multilines [[GH-212](https://github.com/cloudwalk/ethgo/issues/212)]
+- feat: `abi` DecodeStruct uses the `abi` tag instead of the default `mapstructure` [[GH-211](https://github.com/cloudwalk/ethgo/issues/211)]
+- feat: Implement `ens` reverse resolver [[GH-210](https://github.com/cloudwalk/ethgo/issues/210)]
+- fix: Jsonrpc eth_getLogs request cannot return string [[GH-209](https://github.com/cloudwalk/ethgo/issues/209)]
+
+# 0.1.3 (13 June, 2022)
+
+- Fix out-of-bounds reading of bytes during ABI decoding [[GH-205](https://github.com/cloudwalk/ethgo/issues/205)]
+- Update `fastrlp` to `59d5dd3` commit to fix a bug on bytes length check [[GH-204](https://github.com/cloudwalk/ethgo/issues/204)]
+- Fix out-of-bounds RLP unmarshal of transactions [[GH-203](https://github.com/cloudwalk/ethgo/issues/203)]
+
+# 0.1.2 (5 May, 2022)
 
 - Update `btcd` library to new `v0.22.1`
 - Add option in `contract` to send transactions with EIP-1559 [[GH-198](https://github.com/cloudwalk/ethgo/issues/198)]